Fix warnings [-Wold-style-cast] and [-Wuseless-cast]

+++ Bloon.cpp
@@ -117,13 +117,13 @@ void Bloon::initialize(Textureloader* textload)
     }
 }
 
-void Bloon::findWay(int x, int y, int postab, Textureloader* textload)
+void Bloon::findWay(unsigned int x, unsigned int y, int postab, Textureloader* textload)
 {
     // If the mao exist
     if (m_carte != "")
     {
         // If the coord exist
-        if ((unsigned int)x < textload->getMap(m_carte).getSize().x && (unsigned int)y < textload->getMap(m_carte).getSize().y)
+        if (x < textload->getMap(m_carte).getSize().x && y < textload->getMap(m_carte).getSize().y)
         {
             m_color = textload->getMap(m_carte).getPixel(x, y);
             // If the color isn't green
