+++ Entity.cpp
@@ -225,5 +225,5 @@ void Entity::drawRange(bool range)
 
 int Entity::getRange()
 {
-    return (float)(m_range.getRadius());
+    return static_cast<int>(m_range.getRadius());
 }
